+++ src/mediarithmics/plugins/computed-field/ComputedFieldBasePlugin.ts
@@ -19,27 +19,55 @@ export interface OnUpdateBatchPluginResponse<S> extends PluginResponse {
   };
 }
 
-export interface BuildResultPluginResponse<D> extends PluginResponse {
+export interface BuildResultPluginResponse<R> extends PluginResponse {
   data: {
-    result: D;
+    result: R;
   };
 }
 
-export interface RequestData<S, D> {
-  state: S;
-  data: D;
+export enum DataType {
+  USER_ACTIVITY,
+  USER_PROFILE,
+  COMPUTED_FIELD,
 }
 
-export interface RequestDataBatch<S, D> {
+export enum Operation {
+  INSERT,
+  DELETE,
+  UPDATE,
+}
+
+export interface BaseUserActivity {}
+export interface BaseUserProfile {}
+export interface BaseComputedField {}
+
+export interface Update {
+  dataType: DataType;
+  operation: Operation;
+  data: BaseUserActivity | BaseUserProfile | BaseComputedField;
+}
+
+export interface RequestData<S> {
   state: S;
-  data: D[];
+  update: Update;
+}
+
+export interface RequestDataBatch<S> {
+  state: S;
+  updates: Update[];
 }
 
 export interface RequestResult<S> {
   state: S;
 }
 
-export abstract class ComputedFieldPlugin<State, Data, Result> extends BasePlugin {
+export abstract class ComputedFieldPlugin<
+  State,
+  Result,
+  UserActivity extends BaseUserActivity,
+  UserProfile extends BaseUserProfile,
+  ComputedField extends BaseComputedField,
+> extends BasePlugin {
   constructor() {
     super();
     this.initUpdateRoute();
@@ -47,13 +75,28 @@ export abstract class ComputedFieldPlugin<State, Data, Result> extends BasePlugi
     this.initBuildResultRoute();
   }
 
-  abstract onUpdate(state: State | null, data: Data): State | null;
+  abstract onUpdateActivity(state: State | null, userActivity: UserActivity): State | null;
+  abstract onUpdateUserProfile(state: State | null, userProfile: UserProfile, operation: Operation): State | null;
+  abstract onUpdateComputedField(state: State | null, computedField: ComputedField): State | null;
 
   abstract buildResult(state: State | null): Result | null;
 
-  private onUpdateBatch(state: State, data: Data[]): State | null {
-    return data.reduce((acc, curr) => {
-      return this.onUpdate(acc, curr);
+  private getUpdateMethod(state: State | null, update: Update): State | null {
+    switch (update.dataType) {
+      case DataType.USER_ACTIVITY:
+        return this.onUpdateActivity(state, update.data as UserActivity);
+      case DataType.USER_PROFILE:
+        return this.onUpdateUserProfile(state, update.data as UserProfile, update.operation);
+      case DataType.COMPUTED_FIELD:
+        return this.onUpdateComputedField(state, update.data as ComputedField);
+      default:
+        return state;
+    }
+  }
+
+  private onUpdateBatch(state: State, updates: Update[]): State | null {
+    return updates.reduce((acc, curr) => {
+      return this.getUpdateMethod(acc, curr);
     }, state);
   }
 
@@ -73,16 +116,21 @@ export abstract class ComputedFieldPlugin<State, Data, Result> extends BasePlugi
     this.app.post(
       '/v1/computed_field/update/single',
       (req: express.Request<unknown, unknown, string>, res: express.Response) => {
-        const body = this.formatRequestData<RequestData<State, Data>>(req);
-        const updatedState = this.onUpdate(body.state, body.data);
-        const pluginResponse: OnUpdatePluginResponse<State | null> = {
-          status: 'ok',
-          data: {
-            state: updatedState,
-          },
-        };
-        const response = this.formatResponse(req, pluginResponse);
-        res.status(200).send(response);
+        try {
+          const body = this.formatRequestData<RequestData<State>>(req);
+          const updatedState = this.getUpdateMethod(body.state, body.update);
+          const pluginResponse: OnUpdatePluginResponse<State | null> = {
+            status: 'ok',
+            data: {
+              state: updatedState,
+            },
+          };
+          const response = this.formatResponse(req, pluginResponse);
+          res.status(200).send(response);
+        } catch (error) {
+          this.logger.error('Something bad happened on single update route', error);
+          return res.status(500).send({ status: 'error', message: `${(error as Error).message}` });
+        }
       },
     );
   }
@@ -91,15 +139,20 @@ export abstract class ComputedFieldPlugin<State, Data, Result> extends BasePlugi
     this.app.post(
       '/v1/computed_field/update/batch',
       (req: express.Request<unknown, unknown, string>, res: express.Response) => {
-        const body = this.formatRequestData<RequestDataBatch<State, Data>>(req);
-        const updatedState = this.onUpdateBatch(body.state, body.data);
-        const pluginResponse: OnUpdatePluginResponse<State | null> = {
-          status: 'ok',
-          data: {
-            state: updatedState,
-          },
-        };
-        res.status(200).send(this.formatResponse(req, pluginResponse));
+        try {
+          const body = this.formatRequestData<RequestDataBatch<State>>(req);
+          const updatedState = this.onUpdateBatch(body.state, body.updates);
+          const pluginResponse: OnUpdatePluginResponse<State | null> = {
+            status: 'ok',
+            data: {
+              state: updatedState,
+            },
+          };
+          res.status(200).send(this.formatResponse(req, pluginResponse));
+        } catch (error) {
+          this.logger.error('Something bad happened on single update route', error);
+          return res.status(500).send({ status: 'error', message: `${(error as Error).message}` });
+        }
       },
     );
   }
@@ -108,15 +161,20 @@ export abstract class ComputedFieldPlugin<State, Data, Result> extends BasePlugi
     this.app.post(
       '/v1/computed_field/build_result',
       (req: express.Request<unknown, unknown, string>, res: express.Response) => {
-        const body = this.formatRequestData<RequestResult<State>>(req);
-        const buildResult = this.buildResult(body.state);
-        const pluginResponse: BuildResultPluginResponse<Result | null> = {
-          status: 'ok',
-          data: {
-            result: buildResult,
-          },
-        };
-        res.status(200).send(this.formatResponse(req, pluginResponse));
+        try {
+          const body = this.formatRequestData<RequestResult<State>>(req);
+          const buildResult = this.buildResult(body.state);
+          const pluginResponse: BuildResultPluginResponse<Result | null> = {
+            status: 'ok',
+            data: {
+              result: buildResult,
+            },
+          };
+          res.status(200).send(this.formatResponse(req, pluginResponse));
+        } catch (error) {
+          this.logger.error('Something bad happened on single update route', error);
+          return res.status(500).send({ status: 'error', message: `${(error as Error).message}` });
+        }
       },
     );
   }